Update README.md
In this project I have created linear algebra functions and have used them to create a minimum variance portfolio using basic financial mathematics. A minimum variance portfolio is essentially the combination of stocks that will minimize the volatility of the overall portfolio. the first challange of this project was downloading the code