Create Data Download
This function will download data from quandl, you would need an access token from quandl to access it. It first creates .txt file that stores all of the ticker names,then it creates a .txt file for the data associated with the given ticker, this includes the opening price, closing price, volume, etc. for everyday day that the company has become public.